refactor: remove wrapper-only performance targets, rename to performance_targets
Remove dead code from wrapper_performance_targets.rs (PERFORMANCE_FLOOR_MS,
BenchmarkResult, log_performance_target_if_violated) which were only used by
the now-deleted synchronous wrapper mode. Keep log_performance_for_checkpoint
which is still used for checkpoint performance tracking.

Rename wrapper_performance_targets.rs to performance_targets.rs since it's
no longer wrapper-specific. Update all import sites:
- src/observability/mod.rs: update module declaration
- src/commands/git_ai_handlers.rs: update import
- tests/integration/repos/test_repo.rs: define local BenchmarkResult
- tests/integration/performance.rs: use BenchmarkResult from test_repo,
  define local PERFORMANCE_FLOOR_MS constant
- tests/integration/wrapper_performance_targets.rs: renamed to
  performance_targets.rs with only checkpoint tests retained
