+++ tests/integration/main.rs
@@ -73,6 +73,7 @@ mod non_utf8_files;
 mod notes_merge_mixed_fanout;
 mod opencode;
 mod performance;
+mod performance_targets;
 mod pi;
 mod post_commit_unit;
 mod pre_commit_unit;
@@ -119,4 +120,3 @@ mod virtual_attribution_merge;
 mod virtual_attribution_unit;
 mod windsurf;
 mod worktrees;
-mod wrapper_performance_targets;
